refactor(profile): move `isTimeRangeDisabled` to `/profile/utils/helpers`
