+++ app/profile/components/common/toggle-stats-provider.tsx
@@ -4,8 +4,8 @@ import { usePathname, useSearchParams } from 'next/navigation'
 import Link from 'next/link'
 
 import type { ProfileSelectProps } from './selects/props'
-import { isTimeRangeDisabled } from './utils'
 
+import { isTimeRangeDisabled } from '@app/profile/utils/helpers'
 import { ToggleGroup, ToggleGroupItem } from '@app/components/ui/toggle-group'
 import { formatSearchParams } from '@app/utils/formatters'
 import { STATS_PROVIDER, VIEW } from '@app/profile/constants'
