+++ app/profile/utils/helpers/is-time-range-disabled.ts
